error displays and minor changes in model to remove decimal value of price

+++ upahar/store/models.py
@@ -24,7 +24,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name=models.CharField(max_length=150)
-    price=models.DecimalField(default=0,decimal_places=2,max_digits=9)
+    price=models.IntegerField(default=0)
     category=models.ForeignKey(Category,on_delete=models.CASCADE,default=1)
     description=models.CharField(max_length=250,default='',blank=True,null=True)
     image=models.ImageField(upload_to='uploads/product/')
